# feat: parallellize radix2 bowers