--- a/dft/src/butterflies.rs
+++ b/dft/src/butterflies.rs
@@ -1,39 +1,6 @@
 use p3_field::Field;
 use p3_matrix::dense::RowMajorMatrixViewMut;
 
-/// DIT butterfly operation on rows.
-#[inline]
-pub(crate) fn dit_butterfly_on_rows<F: Field>(row_1: &mut [F], row_2: &mut [F], twiddle: F) {
-    let ((prefix_1, shorts_1, suffix_1), (prefix_2, shorts_2, suffix_2)) = unsafe {
-        (
-            row_1.align_to_mut::<F::Packing>(),
-            row_2.align_to_mut::<F::Packing>(),
-        )
-    };
-
-    for (x_1, x_2) in prefix_1.iter_mut().zip(prefix_2) {
-        let x_2_twiddle = *x_2 * twiddle;
-        let sum = *x_1 + x_2_twiddle;
-        let diff = *x_1 - x_2_twiddle;
-        *x_1 = sum;
-        *x_2 = diff;
-    }
-    for (x_1, x_2) in shorts_1.iter_mut().zip(shorts_2) {
-        let x_2_twiddle = *x_2 * twiddle;
-        let sum = *x_1 + x_2_twiddle;
-        let diff = *x_1 - x_2_twiddle;
-        *x_1 = sum;
-        *x_2 = diff;
-    }
-    for (x_1, x_2) in suffix_1.iter_mut().zip(suffix_2) {
-        let x_2_twiddle = *x_2 * twiddle;
-        let sum = *x_1 + x_2_twiddle;
-        let diff = *x_1 - x_2_twiddle;
-        *x_1 = sum;
-        *x_2 = diff;
-    }
-}
-
 /// DIF butterfly operation on rows.
 #[inline]
 pub(crate) fn dif_butterfly_on_rows<F: Field>(row_1: &mut [F], row_2: &mut [F], twiddle: F) {
@@ -94,6 +61,20 @@ pub(crate) fn twiddle_free_butterfly_on_rows<F: Field>(row_1: &mut [F], row_2: &
     }
 }
 
+/// DIT butterfly operation on rows.
+#[inline]
+pub(crate) fn dit_butterfly_on_rows<F: Field>(row_1: &mut [F], row_2: &mut [F], twiddle: F) {
+    let ((prefix_1, shorts_1, suffix_1), (prefix_2, shorts_2, suffix_2)) = unsafe {
+        (
+            row_1.align_to_mut::<F::Packing>(),
+            row_2.align_to_mut::<F::Packing>(),
+        )
+    };
+    dit(
+        prefix_1, shorts_1, suffix_1, prefix_2, shorts_2, suffix_2, twiddle,
+    )
+}
+
 /// DIT butterfly.
 #[inline]
 pub(crate) fn dit_butterfly<F: Field>(
@@ -104,7 +85,20 @@ pub(crate) fn dit_butterfly<F: Field>(
 ) {
     let ((prefix_1, shorts_1, suffix_1), (prefix_2, shorts_2, suffix_2)) =
         mat.packing_aligned_rows(row_1, row_2);
+    dit(
+        prefix_1, shorts_1, suffix_1, prefix_2, shorts_2, suffix_2, twiddle,
+    )
+}
 
+fn dit<F: Field>(
+    prefix_1: &mut [F],
+    shorts_1: &mut [<F as Field>::Packing],
+    suffix_1: &mut [F],
+    prefix_2: &mut [F],
+    shorts_2: &mut [<F as Field>::Packing],
+    suffix_2: &mut [F],
+    twiddle: F,
+) {
     for (x_1, x_2) in prefix_1.iter_mut().zip(prefix_2) {
         let x_2_twiddle = *x_2 * twiddle;
         let sum = *x_1 + x_2_twiddle;