This PR enhances the performance of DFT radix2 bowers by parallelizing the dit and dif butterfly methods across matrix rows in row-major form. Benchmarks indicate above 3x speed improvements. For instance, 

| Bench Name | Previous Execution Time (ms) | New Execution Time (ms) | Improvement Factor |
|------------|------------------------------|-------------------------|--------------------|
| `fft::<p3_goldilocks::Goldilocks, p3_dft::radix_2_dit::Radix2Dit, 100>/65536` | ~168 | 50.986 | ~3.3x |
| `fft::<p3_baby_bear::baby_bear::BabyBear, p3_dft::radix_2_dit::Radix2Dit, 100>/262144` | ~101 | 50.808 | ~2 |
| `m31_fft::<p3_dft::radix_2_dit::Radix2Dit, 100>/262144` | ~800 | 259.91 | ~3.13 |


However, our methods applied to the logic in `radix_2_dit_parallel.rs` exhibited increased execution times, likely due to parallelization overheads; thus, no refactoring was applied to it.