@dlubarov 2023-10-09
Do you have a sense of the performance of the improved `Radix2Dit` (or `Radix2Bowers`, should be similar) compared to `Radix2DitParallel`? Just tried on my M1 and `Radix2DitParallel` seems a little faster with `BabyBear`, about the same with `Goldilocks`. It might vary a lot on different machines though.

I think the nice thing about `Radix2DitParallel` is its parallelization strategy should lead to most accesses hitting L1 cache (depending on L1 size and problem size), but the downside is we pay a price for bit-reversals to maintain that locality.