@jorgeantonio21 2023-10-10
To answer to your previous comment. I am running also on a M1, 16gbs RAM. I am having similar benches to both `Radix2Dit` and `Radix2DitParallel`, for Goldilocks. For `BabyBear`, it is also slightly faster on my machine with `Radix2DitParallel` vs `Radix2Dit`.

I am convinced, with binary exchange, we can improve the performance of `Radix2DitParallel` with a simple approach, as the one used in this PR, but I am still to properly test it. I want to further test with transpose methods.